Added thermal and exozodiacal backgrounds, other changes
Instrument thermal and exozodiacal backgrounds now taken into account.  Exozodi calculated using modern Solar System Haystacks simulation (https://asd.gsfc.nasa.gov/projects/haystacks/haystacks.html), 60 deg inclination.  All spectra now converted to photon units as a beginning step.  Form of final reduced spectrum changed along with noise model.  Some other general cleanup as well.
