+++ SunEarth_4m.init
@@ -1,31 +1,35 @@
 wav_min: 		1.3				# in um
-wav_max: 		1.5				# in um
+wav_max: 		1.5 			# in um
 t_exp: 			144e4					# in second
 obj_tag: 		SunEarth_4m_H2O			# appear in saved pkl file
 st_spec_path: 		template.star.fits			# spec fits file for star
 st_template_spec_path:	lte058.0-4.5-0.0a+0.0.BT-Settl.spec.fits# template spec fits file to generate star fits file
-st_inclination_deg: 	50.0					# in degree
+st_inclination_deg: 	60.0					# in degree
 st_rotation_vel: 	2e3					# in m/s
 st_radial_vel: 		0e3					# in m/s positive is blue shift
 pl_spec_path: 		template.planet.fits			# spec fits file for planet 
 pl_template_spec_path:	GeometricA_Earth_NoCloud_UltraRes.dat, GeometricA_Earth_HighCloud_UltraRes.dat # template spec fits file to generate planet fits file
-pl_inclination_deg: 	50.0					# in degree
+pl_inclination_deg: 	60.0					# in degree
 pl_rotation_vel: 	5e2					# in m/s
 pl_radial_vel: 		20.4e3					# in m/s positive is blue shift
 pl_st_distance:		1.496e11				# in m 1 AU = 1.496e11 m
 pl_radius:		6378000.0				# in m 1 Earth Radius = 6378000.0 m
 pl_illumination:	0.5					# fraction of illumination for a planet
-spec_reso: 		128e2				# spectral resolution
+zodi_spec_path:		modern_cube_zodi1inc60dist10_1.32-1.63um.fits	#Exozodi simulation fits file
+spec_reso: 		300				# spectral resolution
 distance: 		5.0					# in pc
 spec_tran_path: 	None					# atmosphere transmission file path
 spec_radi_path: 	None					# atmosphere emission file path
-telescope_size: 	12.0					# in m
-pl_st_contrast: 	5e-8			# star light suppression at fiber position
-read_noise: 		0.0					# in electron / pixel per read
-dark_current: 		0e0					# in elsctron / s / pixel
+telescope_size: 	4.0					# in m
+num_surfaces:		10  			# number of warm optical surfaces in the telescope
+temperature:		265				# temperature of the telescope in K
+exozodi_level:		1.0				# level of exozodi relative to the Solar System
+pl_st_contrast: 	1e-10			# star light suppression at fiber position
+read_noise: 		2.0					# in electron / pixel per read
+dark_current: 		0.002					# in elsctron / s / pixel
 fiber_size: 		1.0					# in lambda / D fiber sky projection 
 pixel_sampling: 	3.0					# pixels per resolution element
-throughput: 		0.1					# end to end throughput
+throughput: 		0.15					# end to end throughput
 wfc_residual: 		5.0					# in nm
 template_path: 		template.H2O.fits			# spec fits file for template
 template_tag: 		H2O				# appear in saved pkl file
