+++ goldrefrac.txt
@@ -1,0 +1,148 @@
+# this file is part of refractiveindex.info database
+# refractiveindex.info database is in the public domain
+# copyright and related rights waived via CC0 1.0
+#
+#REFERENCES: "K. M. McPeak, S. V. Jayanti, S. J. P. Kress, S. Meyer, S. Iotti, A. Rossinelli, and D. J. Norris. Plasmonic films can easily be better: Rules and recipes, <a href=\"https://doi.org/10.1021/ph5004237\"><i>ACS Photonics</i> <b>2</b>, 326-333 (2015)</a>"
+#COMMENTS: "Thermally evaporated Au at 10 Ã…/s with base pressure of 3e-8 Torr, template stripped from Si wafer"
+#
+0.3 1.699838715 1.973157788
+0.31 1.746948702 1.974499878
+0.32 1.782941044 1.959795593
+0.33 1.804652963 1.931950392
+0.34 1.804621334 1.896796816
+0.35 1.763712432 1.870684779
+0.36 1.713434851 1.887670254
+0.37 1.689202065 1.922090949
+0.38 1.678059067 1.948381952
+0.39 1.672504217 1.966512231
+0.4 1.665616091 1.973924254
+0.41 1.653911114 1.975227068
+0.42 1.63743714 1.970304643
+0.43 1.612343317 1.959477219
+0.44 1.581523054 1.940313163
+0.45 1.538326354 1.910745397
+0.46 1.475285274 1.872315849
+0.47 1.386263242 1.827683172
+0.48 1.252905499 1.782097694
+0.49 1.064360219 1.76786953
+0.5 0.848474841 1.828280492
+0.51 0.661635502 1.964525779
+0.52 0.52912664 2.129735899
+0.53 0.438041087 2.294990195
+0.54 0.372502061 2.451664289
+0.55 0.323930966 2.597158307
+0.56 0.284960267 2.738978341
+0.57 0.254069233 2.871855006
+0.58 0.228936734 2.999518633
+0.59 0.207122197 3.1213458
+0.6 0.188789629 3.241703491
+0.61 0.172726107 3.356759495
+0.62 0.160071695 3.465634566
+0.63 0.14636792 3.577628204
+0.64 0.135417727 3.685964997
+0.65 0.125499436 3.792327795
+0.66 0.117832095 3.896137113
+0.67 0.111117685 3.999543366
+0.68 0.105744888 4.102704228
+0.69 0.101509306 4.205033191
+0.7 0.098607898 4.305081128
+0.71 0.097790271 4.403812319
+0.72 0.098623269 4.499191766
+0.73 0.096903666 4.59580138
+0.74 0.099463321 4.687205239
+0.75 0.098537395 4.78107829
+0.76 0.099206649 4.872354868
+0.77 0.101706737 4.958865219
+0.78 0.101315225 5.051560628
+0.79 0.103067755 5.148069829
+0.8 0.10422723 5.223682926
+0.81 0.105937788 5.313023886
+0.82 0.104338656 5.407391844
+0.83 0.104534544 5.49244276
+0.84 0.107001078 5.577046641
+0.85 0.111012574 5.659268839
+0.86 0.11076329 5.749197205
+0.87 0.112038779 5.830570529
+0.88 0.114173034 5.912954886
+0.89 0.116004573 5.992867182
+0.9 0.116828465 6.081993825
+0.91 0.116997234 6.163820922
+0.92 0.11945701 6.241575921
+0.93 0.120635837 6.328550624
+0.94 0.122558964 6.410383818
+0.95 0.124002721 6.491793025
+0.96 0.123935366 6.573991175
+0.97 0.126244253 6.662877577
+0.98 0.127974627 6.73805443
+0.99 0.129942772 6.819155749
+1 0.132311724 6.904527963
+1.01 0.133907614 6.978318655
+1.02 0.135401137 7.061609835
+1.03 0.138148809 7.141574413
+1.04 0.137937636 7.222466808
+1.05 0.139246667 7.302149658
+1.06 0.134049061 7.406616579
+1.07 0.135717763 7.48280825
+1.08 0.140363062 7.556831465
+1.09 0.138686104 7.637030433
+1.1 0.138418628 7.719336741
+1.11 0.139442944 7.79996438
+1.12 0.141677831 7.874202982
+1.13 0.141308241 7.960965269
+1.14 0.140235564 8.050382973
+1.15 0.142817756 8.114887363
+1.16 0.146729336 8.201495565
+1.17 0.152747848 8.26754691
+1.18 0.149693183 8.350413645
+1.19 0.147906483 8.443510901
+1.2 0.149395088 8.515005513
+1.21 0.151984867 8.589999965
+1.22 0.153423422 8.666864413
+1.23 0.153670698 8.744347585
+1.24 0.157748208 8.81753279
+1.25 0.159662147 8.900982642
+1.26 0.165494236 8.987624177
+1.27 0.159382324 9.068759713
+1.28 0.163849184 9.146521008
+1.29 0.166211578 9.226192405
+1.3 0.166156462 9.303580384
+1.31 0.170115488 9.372456417
+1.32 0.169795214 9.445201449
+1.33 0.174809869 9.530926424
+1.34 0.176209118 9.61244244
+1.35 0.179585571 9.687582308
+1.36 0.177483381 9.769365402
+1.37 0.184295562 9.839357949
+1.38 0.186297714 9.916385775
+1.39 0.190837556 9.992267959
+1.4 0.191102283 10.07156989
+1.41 0.194361931 10.14138928
+1.42 0.205426553 10.22263176
+1.43 0.208361137 10.32150252
+1.44 0.206557078 10.37509835
+1.45 0.208044 10.46581494
+1.46 0.205758558 10.54050931
+1.47 0.212405614 10.6421387
+1.48 0.216697642 10.70593097
+1.49 0.219902453 10.78046182
+1.5 0.215398722 10.85893165
+1.51 0.21956893 10.91137986
+1.52 0.222004156 11.00587506
+1.53 0.230589098 11.08617027
+1.54 0.235210701 11.16849695
+1.55 0.238229889 11.26307034
+1.56 0.236761828 11.31706924
+1.57 0.235275342 11.38092063
+1.58 0.240819258 11.44281407
+1.59 0.246290147 11.53172402
+1.6 0.243652913 11.64299647
+1.61 0.25211538 11.70912303
+1.62 0.248830057 11.78354431
+1.63 0.264448459 11.86507198
+1.64 0.258451056 11.92798378
+1.65 0.265882655 12.01168987
+1.66 0.278561249 12.06969744
+1.67 0.278188678 12.1875916
+1.68 0.28090035 12.2404618
+1.69 0.290244511 12.34683125
+1.7 0.316481841 12.37053599
