+++ hci.sh
@@ -1,0 +1,3 @@
+python2 create_planet_template_spec.py
+python2 hci_hrs_sim.py
+
